project1/221300079_王俊童/requirements.txt
numpy>=1.19.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
jieba>=0.42.0
tqdm>=4.48.0
python-Levenshtein>=0.12.0
transformers>=4.0.0
torch>=1.7.0
pandas>=1.1.0
tensorboard>=2.4.0
protobuf>=3.12.0
pypinyin>=0.40.0
opencc>=1.1.1